Move genAI object to objects section
